requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
boost=1.68.0=py36h8619c78_1001
boost-cpp=1.68.0=h11c811c_1000
descriptastorus=2.2.0=py_0
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
pandas=0.25.0=py36hb3f55d8_0
python=3.6.8=h0371630_0
pytorch=1.1.0=py3.6_cuda9.0.176_cudnn7.5.1_0
tensorboard=1.13.1=py36_0
torchvision=0.3.0=py36_cu9.0.176_1
rdkit=2019.03.4.0=py36hc20afe1_1
readline=7.0=h7b6447c_5
scikit-learn=0.21.2=py36hcdab131_1
scipy=1.3.0=py36h921218d_1
tqdm=4.32.1=py_0
typing=3.6.4=py36_0